[et-trace] [SW-12448] Fix build warnings and add option to enable warnings as errors

+++ et-trace/tests/common/user_args.c
@@ -15,7 +15,7 @@ void parse_args(int argc, const char **argv, struct user_args *uargs)
     while ((ret = getopt(argc, (char *const *)argv, ":s:o:h")) != -1) {
         switch (ret) {
         case 's':
-            uargs->seed = atoi(optarg);
+            uargs->seed = (unsigned int)atoi(optarg);
             break;
         case 'o':
             uargs->output = optarg;
